move prover_to_json to its own Rust module
